Move import of SatpyProduct and SatpyScene to downsat.satpy
