+++ downsat/satpy/__init__.py
@@ -1,0 +1,7 @@
+from downsat.data_sources.satpy import SatpyProduct, SatpyScene
+
+
+__all__ = [
+    "SatpyProduct",
+    "SatpyScene",
+]
